refactor: massive 2-day overhaul and restructuring of core modules, setup, and documentation
