+++ README.md
@@ -1,0 +1,43 @@
+# cayman_office_system
+
+## Overview
+
+An integrated platform for managing offshore funds based in the Cayman Islands, tailored for LIFE Asset Management.
+
+### Key Features
+- Order processing
+- Trade execution
+- Account management
+- Comprehensive reporting
+
+## Installation
+
+```bash
+pip install .
+```
+
+## Usage
+
+```bash
+python -m cayman_office_system
+```
+
+## Requirements
+- Python >= 3.8
+- See requirements.txt for details
+
+## Author
+
+**June Young Park**  
+AI Management Development Team Lead & Quant Strategist at LIFE Asset Management
+
+LIFE Asset Management is a hedge fund management firm that integrates value investing and engagement strategies with quantitative approaches and financial technology, headquartered in Seoul, South Korea.
+
+## Contact
+
+- Email: juneyoungpaak@gmail.com
+- Location: TWO IFC, Yeouido, Seoul
+
+---
+
+For more information, please contact the Development Team at LIFE Asset Management.
