+++ cayman_office_system/__main__.py
@@ -1,4 +1,5 @@
-# cayman_office_system/__main__.py
-if __name__ == "__main__":
-    print("|| (LIFE_DEV_MODULE) cayman_office_system ||")
+from cayman_office_system.structure_generator import generate_data_folder_structure
 
+if __name__ == "__main__":
+    generate_data_folder_structure()
+    print("| cayman_office_system initialized.")
