+++ cayman_office_system/cos_dataset_loader/dataset_constants.py
@@ -41,9 +41,3 @@ MAPPING_COLUMNS_ORDER = {
     'Symbol': 'ticker',
     'Amount': 'amount'
 }
-
-MAPPING_SECTOR = {
-    'NAME': 'name',
-    'GICS_SECTOR_NAME': 'sector',
-    'ticker_bbg_index': 'market_index'
-}
