+++ cayman_office_system/dataset_account_controller/timeseries_account_generator.py
@@ -3,7 +3,7 @@ from .dataset_operation_parser import get_sum_of_revenues_and_expenses_at_month_
 from .valid_account_logs_consts import VALID_ACCOUNT_LOGS, VALID_DATES_FOR_INCOME_LOSS_DATA
 from .flow_utils import get_timeseries_flow
 from shining_pebbles import get_today, get_date_range
-from ..dataset_loader import get_df_usdkrw, get_usdkrw_of_date
+from cayman_office_system.market_database import get_df_usdkrw, get_usdkrw_of_date
 import pandas as pd
