--- cayman_office_system/docs/LKEF_INFO.md
@@ -1,17 +1,0 @@
-# LKEF INCEPTION DATE
-
-2024-09-02
-
-# LKEF NAV
-
-## 2024-11-30
-
-15,856,996.72
-
-## 2024-10-31
-
-15,677,459.74
-
-## 2024-09-30
-
-16,086,961.37
