--- cayman_office_system/docs/README.md
@@ -1,26 +1,0 @@
-# cayman_office_system
-
-## Overview
-
-An integrated platform for managing offshore funds based in the Cayman Islands, tailored for LIFE Asset Management.
-
-### Key Features
-
-- Order processing
-- Trade execution
-- Account management
-- Comprehensive reporting
-
-## Development
-
-This system was independently developed by June Young Park, DEV TEAM LEAD at LIFE Asset Management.
-
-### Technical Highlights
-
-- Custom-built from the ground up
-- Optimized for LIFE Asset Management's specific workflows
-- Integrates seamlessly with existing financial systems
-
----
-
-For more information, please contact the Development Team at LIFE Asset Management.
