+++ cayman_office_system/market_database/market_information.py
@@ -1,7 +1,6 @@
 from  shining_pebbles import get_today
-from .dataset_loader import get_df_sector_ks, get_df_price_usdlevetf, get_df_cap_usdlevetf
+from .pseudo_database import get_df_sector_ks, get_df_price_usdlevetf, get_df_cap_usdlevetf
 from .birdeye_connector import get_df_prices_of_ks_stock
-from .finance_utils import get_ticker_from_ticker_bbg, get_ticker_bbg_of_ticker
 
 
 ROW_INFO_FOR_USD_LEVERAGED_EFT = {
@@ -15,7 +14,7 @@ ROW_INFO_FOR_USD_LEVERAGED_EFT = {
 }
 
 
-class MarketInfo:
+class MarketInformation:
     def __init__(self, name='ks'):
         self.name = name
         self.prices = self.get_prices()
@@ -55,34 +54,3 @@ class MarketInfo:
         self.mapping_sector = self.sector['sector'].to_dict()
         return self.mapping_sector
     
-
-def get_mapping_ks_name():
-    name_sector = get_df_sector_ks()
-    return name_sector['name'].to_dict()
-
-def get_mapping_ks_sector():
-    name_sector = get_df_sector_ks()
-    return name_sector['sector'].to_dict()
-
-def get_ks_market_info():
-    return MarketInfo(name='ks').info
-
-def append_market_info_to_df(df):
-    df['ticker_bbg'] = df['ticker'].map(lambda x: get_ticker_bbg_of_ticker(x))
-    ks = get_ks_market_info().reset_index()
-    df = df.merge(ks, how='left', on='ticker_bbg')
-    return df
-
-def get_ks_equity_info(tickers):
-    tickers_bbg = [get_ticker_bbg_of_ticker(ticker) for ticker in tickers]
-    ks = get_ks_market_info()
-    df = ks[ks.index.isin(tickers_bbg)]
-    df.columns.name = get_today()
-    return df
-
-def get_list_of_ks_stocks():
-    stock_info = get_ks_market_info()
-    stock_info['ticker'] = stock_info.index.map(get_ticker_from_ticker_bbg)
-    stock_info = stock_info.fillna('-')
-    return stock_info
-
