+++ cayman_office_system/market_database/pseudo_database_consts.py
@@ -1,0 +1,6 @@
+
+MAPPING_SECTOR = {
+    'NAME': 'name',
+    'GICS_SECTOR_NAME': 'sector',
+    'ticker_bbg_index': 'market_index'
+}
