+++ cayman_office_system/order/order_consts.py
@@ -1,0 +1,4 @@
+
+COLUMNS_INFO_ORDER = ['date', 'ticker', 'name_kr', 'name', 'amount', 'price_acquisition_approx', 'price_last', 'return_approx', 'acquisition_approx_krw', 'evaluation_krw', 'acquisition_approx_usd', 'evaluation_usd', 'cap(/1e8)', 'sector', 'market_index']
+COLUMNS_HOLDINGS = ['ticker', 'name_kr', 'name', 'amount', 'price_last', 'evaluation_krw', 'evaluation_usd', 'sector']
+
