+++ cayman_office_system/path_director.py
@@ -1,7 +1,7 @@
 import os
 
-ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-BASE_DIR = os.path.join(ROOT_DIR)
+# ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
+BASE_DIR = os.path.join('data-cos')
 
 DIR_SECTOR = 'dataset-sector'
 DIR_MARKET = 'dataset-market'
