+++ cayman_office_system/trades/__init__.py
@@ -1,0 +1,2 @@
+from .trades import *
+from .trades_synthetic import *
