+++ cayman_office_system/trades/stock/__init__.py
@@ -1,0 +1,1 @@
+from .stock import *
