+++ cayman_office_system/trades/trade/transaction/transaction_consts.py
@@ -1,0 +1,13 @@
+
+KEYS_TRANSACTION = ['Type',
+ 'ISIN Code / Abbr. Code',
+ 'Security Description',
+ 'Total No. of Shares',
+ 'Average Price',
+ 'Considerations',
+ 'Commission',
+ 'Sales Tax',
+ 'Capital Gains Tax',
+ 'Net Amount']
+
+KEY_SELLBUY = 'No. of Shares / Price'
